# Public/openapi.yaml
openapi: 3.0.1
info:
  title: Synthstaff API
  version: "1.0.0"
  description: API for managing AI staffing requests and retrieving matched AI employees.
servers:
  - url: https://synthstaff-api.onrender.com
paths:
  /api/v1/staffing-request:
    post:
      operationId: submitStaffingRequest
      summary: Submit a staffing request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role_title:
                  type: string
                skills_required:
                  type: array
                  items:
                    type: string
                urgency_level:
                  type: string
                remote_allowed:
                  type: boolean
                additional_notes:
                  type: string
      responses:
        '201':
          description: Staffing request received
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: object
  /api/v1/matched-candidates:
    get:
      operationId: getMatchedCandidates
      summary: Get matched candidates by request ID
      parameters:
        - in: query
          name: request_id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: A list of matched candidates
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  candidates:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        skills:
                          type: array
                          items:
                            type: string